Radiation

Core Radiation Functions

ClearSky.planck — Function

planck(ν, T)

Compute black body intensity [W/m$^2$/cm$^{-1}$/sr] using Planck's law

Arguments

ν: wavenumger [cm$^{-1}$]

T: temperature [Kelvin]

ClearSky.normplanck — Function

normplanck(ν, T)

Compute black body intensity [W/m$^2$/cm$^{-1}$/sr] using Planck's law, normalized by the power emitted per unit area at the given temperature (stefanboltzmann),

B(ν,T)/σT^4

yielding units of 1/cm$^{-1}$/sr.

Arguments

ν: wavenumger [cm$^{-1}$]

T: temperature [Kelvin]

ClearSky.stefanboltzmann — Function

stefanboltzmann(T)

Compute black body radiation power using the Stefan-Boltzmann law, $σT^4$ [W/m$^2$].

ClearSky.equilibriumtemperature — Function

equilibriumtemperature(F, A)

Compute the planetary equilibrium temperature, or equivalent blackbody temperature of a planet.

$(\frac{(1 - A)F}{4\sigma})^{1/4}$

Arguments

F: stellar flux [W/m$^2$]

A: albedo

equilibriumtemperature(L, A, R)

Compute the planetary equilibrium temperature, or equivalent blackbody temperature of a planet.

$(\frac{(1 - A)L}{16 \sigma \pi R^2})^{1/4}$

Arguments

L: stellar luminosity [W]

A: albedo

R: orbital distance [m]

ClearSky.dτdP — Function

dτdP(σ, g, μ)

Evaluate the differential increase in optical depth in pressure coordinates, equivalent to the schwarzschild equation without Planck emission.

$\frac{dτ}{dP} = σ\frac{\textrm{N}_A}{g μ}$

where $N_A$ is Avogadro's number.

Arguments

σ: absorption cross-section [cm$^2$/molecule]

g: gravitational acceleration [m/s$^2$]

μ: mean molar mass [kg/mole]

ClearSky.schwarzschild — Function

schwarzschild(I, ν, σ, T, P)

Evaluate the Schwarzschild differential equation for radiative transfer with units of length/height [m] and assuming the ideal gas law.

$\frac{dI}{dz} = σ\frac{P}{k_B T}[B_ν(T) - I]$

where $B_ν$ is planck's law.

Arguments

I: radiative intensity [W/m$^2$/cm$^{-1}$/sr]

ν: radiation wavenumber [cm$^{-1}$]

σ: absorption cross-section [cm$^2$/molecule]

T: temperature [K]

P: pressure [Pa]

schwarzschild(I, ν, σ, g, μ, T)

Evaluate the Schwarzschild differential equation for radiative transfer with pressure units [Pa] and assuming the ideal gas law.

$\frac{dI}{dP} = σ\frac{\textrm{N}_A}{g μ}[B_ν(T) - I]$

where $B_ν$ is planck's law and $N_A$ is Avogadro's number.

Arguments

I: radiative intensity [W/m$^2$/cm$^{-1}$/sr]

ν: radiation wavenumber [cm$^{-1}$]

σ: absorption cross-section [cm$^2$/molecule]

g: gravitational acceleration [m/s$^2$]

μ: mean molar mass [kg/mole]

T: temperature [K]

Radiation Unit Conversions

ClearSky.ν2f — Function

Convert wavenumber [cm$^{-1}$] to frequency [1/s]

ClearSky.f2ν — Function

Convert frequency [1/s] to wavenumber [cm$^{-1}$]

ClearSky.ν2λ — Function

Convert wavenumber [cm$^{-1}$] to wavelength [m]

ClearSky.λ2ν — Function

Convert wavelength [m] to wavenumber [cm$^{-1}$]

ClearSky.λ2f — Function

Convert wavelength [m] to frequency [1/s]

ClearSky.f2λ — Function

Convert frequency [1/s] to wavelength [m]