Modeling

ClearSky.opticaldepth — Function

opticaldepth(P₁, P₂, g, fT, fμ, θ, absorbers...; tol=1e-5)

Compute the optical depth (τ) between two pressure levels

Arguments

P₁: first pressure level [Pa]

P₂: second pressure level [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P). This may be any callable object, like MoistAdiabat, for example.

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

θ: angle [rad] of path, must be ∈ [0,π/2), where 0 is straight up

absorbers: at least one gas object and any number of CIATables and functions in the form σ(ν, T, P)

Returns a vector of optical depths across all wavenumbers stored in gas objects. The tol keyword argument adjusts integrator error tolerance.

ClearSky.transmittance — Function

transmittance(τ)

Evaluate transmittance from optical depth, $t = e^{-τ}$

transmittance(P₁, P₂, g, fT, fμ, θ, absorbers...; tol=1e-5)

Compute the transmittance between two pressure levels

Arguments

P₁: first pressure level [Pa]

P₂: second pressure level [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P). This may be any callable object, like MoistAdiabat, for example.

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

θ: angle [rad] of path, must be ∈ [0,π/2), where 0 is straight up

absorbers: at least one gas object and any number of CIATables and functions in the form σ(ν, T, P)

Returns a vector of transmittances across all wavenumbers stored in gas objects. The tol keyword argument adjusts integrator error tolerance.

ClearSky.outgoing — Function

outgoing(Pₛ, Pₜ, g, fT, fμ, absorbers; nstream=3, tol=1e-5)

Compute outgoing radiative flux [W/m$^2$/cm$^{-1}$] at each wavenumber defined in the provided gas object(s). Total flux [W/m$^2$] can be evaluated with the trapz function.

Arguments

Pₛ: surface pressure [Pa]

Pₜ: top of atmopshere pressure [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P). This may be any callable object, like MoistAdiabat, for example.

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

absorbers: at least one gas object and any number of CIATables and functions in the form σ(ν, T, P)

The keyword argument nstream specifies how many independent streams, or beam angles through the atmosphere, to integrate. The keyword argument tol is a numerical error tolerance passed to the radau integrator.

ClearSky.incoming — Function

incoming(Pₜ, Pₛ, g, fT, fμ, absorbers; nstream=3, tol=1e-5)

Compute incoming radiative flux [W/m$^2$/cm$^{-1}$] at each wavenumber defined in the provided gas object(s). Total flux [W/m$^2$] can be evaluated with the trapz function.

Arguments

Pₜ: top of atmopshere pressure [Pa]

Pₛ: surface pressure [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P). This may be any callable object, like MoistAdiabat, for example.

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

absorbers: at least one gas object and any number of CIATables and functions in the form σ(ν, T, P)

The keyword argument nstream specifies how many independent streams, or beam angles through the atmosphere, to integrate. The keyword argument tol is a numerical error tolerance passed to the radau integrator.

ClearSky.trapz — Function

trapz(x, y)

Integrate a sorted group of coordinates using the composite trapezoidal rule.