Modeling

ClearSky.opticaldepth — Function

opticaldepth(P₁, P₂, g, fT, fμ, θ, absorbers...; tol=1e-5)

Compute the optical depth (τ) between two pressure levels

Arguments

P₁: first pressure level [Pa]

P₂: second pressure level [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P)

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

θ: angle [rad] of path, must be ∈ [0,π/2)

absorbers: any number of gas objects, CIATables, and functions in the form σ(ν, T, P)

Returns a vector of optical depths across all wavenumbers stored in gas objects. The tol keyword argument adjusts integrator error tolerance.

ClearSky.transmittance — Function

transmittance(τ)

Evaluate transmittance from optical depth, $t = e^{-τ}$

transmittance(P₁, P₂, g, fT, fμ, θ, absorbers...; tol=1e-5)

Compute the transmittance between two pressure levels

Arguments

P₁: first pressure level [Pa]

P₂: second pressure level [Pa]

g: gravitational acceleration [m/s$^2$]

fT: temperature [K] as a function of pressure [Pa], fT(P)

fμ: mean molar mass as a function of temperature [K] and pressure [Pa], fμ(T,P)

θ: angle [rad] of path, must be ∈ [0,π/2)

absorbers: any number of gas objects, CIATables, and functions in the form σ(ν, T, P)

Returns a vector of transmittances across all wavenumbers stored in gas objects. The tol keyword argument adjusts integrator error tolerance.