Orbits and Insolation

A collection of functions are available for working with general elliptical orbits and insolation patterns for planets orbiting stars.

Function arguments are defined below

Argument	Definition	Units

a	semi-major axis	m

e	eccentricity	dimensionless

E	eccentric anomaly	radians

f	true anomaly or stellar longitude	radians

γ	obliquity	radians

m	star mass	kg

p	precession angle	radians

ϕ	latitude	radians

ϕₛ	substellar latitude	radians

rₐ	apoapsis distance	m

rₚ	periapsis distance	m

t	time	seconds

T	orbital period	seconds

The mass m, most precisely, should be the sum of the star mass and planet mass, $m_s + m_p$. For most cases the planet mass is negligible, however, and $m_s + m_p \approx m_s$.

The precession angle p is defined so that when $p=0$, the northern hemisphere is tilted directly toward the star at periapsis. This means that northern summer occurs when planet is closet to the star. Different values of $p ∈ [0,2π]$ control when in the orbital path the equinoxes and solstices occur.

ClearSky.apoapsis — Method

apoapsis(a, e)

Compute the apoapsis (farthest distance) distance using semi-major axis and eccentricity

ClearSky.eccentricanomaly — Method

eccentricanomaly(t, a, m, e)

Numerically compute the eccentric anomaly using Kepler's equation

ClearSky.eccentricity — Method

eccentricity(rₚ, rₐ)

Compute eccentricity

ClearSky.meananomaly — Method

meananomaly(E, e)

Compute the mean anomaly

ClearSky.orbit — Function

orbit(a, m, e, N=1000)

Create a distance time-series of N points for an elliptical orbit

ClearSky.orbitaldistance — Method

orbitaldistance(t, a, m, e)

Compute the distance of a planet from its host

ClearSky.orbitaldistance — Method

orbitaldistance(a, f, e)

Compute the distance of a planet from its host

ClearSky.orbitalperiod — Method

orbitalperiod(a, m)

Kepler's Third Law describing the orbital period of an elliptical orbit

ClearSky.periapsis — Method

periapsis(a, e)

Compute the periapsis (closest approach) distance using semi-major axis and eccentricity

ClearSky.semimajoraxis — Method

semimajoraxis(T, m)

Compute the semi-major axis of an orbit

ClearSky.trueanomaly — Method

trueanomaly(t, a, m, e)

Compute the true anomaly

ClearSky.trueanomaly — Method

trueanomaly(E, e)

Compute the true anomaly

ClearSky.annualfluxfactor — Method

annualfluxfactor(a, m, e, ϕ, γ, p)

Compute the annually averaged flux factor for a latitude ϕ on a planet in a general elliptical orbit

ClearSky.annualfluxfactors — Method

annualfluxfactor(a, m, e, ϕ, γ, p; nϕ=181)

Compute a range of annually averaged flux factors for a planet in a general elliptical orbit. Returns a latitude vector (row values) and a vector of flux factors. nϕ indicates the number of latitude samples.

ClearSky.diurnalfluxfactor — Method

diurnalfluxfactor(t, a, m, e, ϕ, γ, p)

Compute the diurnally averaged fraction of incoming stellar flux received by a point at latitude ϕ for a general elliptical orbit

ClearSky.diurnalfluxfactor — Method

diurnalfluxfactor(ϕ, f, γ)

Compute the diurnally averaged fraction of incoming stellar flux received by a point at latitude ϕ when the planet is at solar longitude (true anomaly) f, with obliquity γ

ClearSky.diurnalfluxfactor — Method

diurnalfluxfactor(ϕ, ϕₛ)

Compute the diurnally averaged fraction of incoming stellar flux received by a point at latitude ϕ when the substellar latitude is ϕₛ

ClearSky.diurnalfluxfactors — Method

diurnalfluxfactors(a, m, e, γ, p; nt=251, nϕ=181)

Compute a grid of diurnally averaged fractions of incoming stellar flux for a planet in a general elliptical orbit. Returns a time vector (column values), latitude vector (row values), and the grid of flux factors. nt indicates the number of time samples around the orbit and nϕ indicates the number of latitudes.

ClearSky.diurnalfluxfactors — Method

diurnalfluxfactors(γ; nf=251, nϕ=181)

Compute a grid of diurnally averaged fractions of incoming stellar flux received by a point at latitude ϕ for a planet with obliquity γ in a circular orbit. Returns a solar longitude vector (column values), latitude vector (row values), and the grid of flux factors. nf indicates the number of points around the orbit and nϕ indicates the number of latitudes.

ClearSky.hourangle — Method

hourangle(ϕ, ϕₛ)

Compute the hour angle

ClearSky.substellarlatitude — Method

substellarlatitude(f, γ)

Compute the latitude of the substellar point for a given solar longitude (true anomaly) and obliquity